requirements.txt
streamlit
pandas
pyarrow
plotly
pillow